Move Base16 builders to ...encoding.base16 pkg + deprecate ...encoding.builders pkg classes/functions
