+++ library/encoding-base16/api/encoding-base16.api
@@ -21,6 +21,24 @@ public final class io/matthewnelson/encoding/base16/Base16$Config : io/matthewne
 	public synthetic fun <init> (ZBZLkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
+public final class io/matthewnelson/encoding/base16/Base16ConfigBuilder {
+	public field encodeToLowercase Z
+	public field isLenient Z
+	public field lineBreakInterval B
+	public fun <init> ()V
+	public fun <init> (Lio/matthewnelson/encoding/base16/Base16$Config;)V
+	public final fun build ()Lio/matthewnelson/encoding/base16/Base16$Config;
+	public final fun strict ()Lio/matthewnelson/encoding/base16/Base16ConfigBuilder;
+}
+
+public final class io/matthewnelson/encoding/base16/BuildersKt {
+	public static final fun Base16 ()Lio/matthewnelson/encoding/base16/Base16;
+	public static final fun Base16 (Lio/matthewnelson/encoding/base16/Base16$Config;Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base16/Base16;
+	public static final fun Base16 (Lkotlin/jvm/functions/Function1;)Lio/matthewnelson/encoding/base16/Base16;
+	public static final fun Base16 (Z)Lio/matthewnelson/encoding/base16/Base16;
+	public static synthetic fun Base16$default (ZILjava/lang/Object;)Lio/matthewnelson/encoding/base16/Base16;
+}
+
 public final class io/matthewnelson/encoding/builders/Base16ConfigBuilder {
 	public field encodeToLowercase Z
 	public field isLenient Z
